Upgraded to Nano every
So more memory don't have to double the display lines and fixed problems with heading indicator.
I previously did not understand how the yaw self corrected to magnetic heading over time and was trying to implement what they already had done. The Bosch folks did it right so I got rid of the search for max flux etc. Now its correct once it says its calibrated but a lot of pitch and yaw will cause it to go out of calibration and then when it comes back into calibration it can be off by 10-20 degrees ... not good. Definitely not to be used for actual heading information.